feat: move worker to separate file
